requirements.txt
accelerate==1.6.0
demjson3==3.0.6
diffusers==0.33.1
fastapi==0.115.12
google-generativeai==0.8.5
huggingface_hub==0.30.2
json_repair==0.41.1
langchain==0.3.24
langchain_core==0.3.55
langchain_ollama==0.3.2
langchain_openai==0.3.14
openai==1.76.0
opencv_python_headless==4.11.0.86
Pillow==11.2.1
protobuf==6.30.2
pytesseract==0.3.13
python-dotenv==1.1.0
python-multipart==0.0.20
Requests==2.32.3
torch==2.6.0
transformers==4.51.3
uvicorn==0.34.2
